Uploading SDD harmonization .Rmd to src folder
I am uploading the SDD harmonization .Rmd to the src folder so that it can be converted to a function once it has been approved.

Pre-processing step: the first section of this file pre-processes the data in the same way that it is processed in the targets pipeline when I was working in it. 

Harmonization step: the second section is where I tried to convert the chl-a harmonization function structure for sdd, but there were issues with this conversion so some adjustments had to be made. Those being:
- General Method Tier instead of Analytical tier:
    - time tag and scope tags were created instead of the tags used in chl-a. These tags were used in determining the tier of the data.
- General Data Flags:
    - field_flag: flags relating to field conditions and whether or not the secchi hit the bottom
    - depth_flag: run as it was made for chl-a
    - mdl_flag: flag based on whether a measurement is above or below 0.01 m.
    - approx_flag: run as it was made for chl-a
    - greater_flag: flag based on '>' or '<' symbols in the data, and if the measurement is greater than 31m.
- Unit conversion was done earlier in the workflow because subsequent steps relied on this conversion.